Testes de unidade ProductService

+++ backend/src/main/java/com/devsuperior/dscatalog/services/ProductService.java
@@ -65,6 +65,9 @@ public class ProductService {
 
     @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
         try {
             repository.deleteById(id);
         } catch (DataIntegrityViolationException e) {
